application error codes can have values up to 2^16-1, so frames using it should parse and handle as long
Thanks to Wesley Rosenblum for reporting this issue.

+++ src/main/java/net/luminis/quic/QuicStream.java
@@ -93,11 +93,11 @@ public interface QuicStream {
      *
      * @param  applicationProtocolErrorCode
      */
-    void closeInput(int applicationProtocolErrorCode);
+    void closeInput(long applicationProtocolErrorCode);
 
     /**
      * Reset the (sending part of the) stream.
      * @param errorCode
      */
-    void resetStream(int errorCode);
+    void resetStream(long errorCode);
 }
